refactor: 💡 cleanup
