+++ packages/ts-ioc-container/lib/features/instanceHook/decorators/IMethodsMetadataCollector.ts
