+++ packages/ts-ioc-container/lib/features/instanceHook/decorators/MethodsMetadataCollector.ts
