+++ packages/ts-ioc-container/lib/index.ts
@@ -10,11 +10,11 @@ export { IInjector } from './core/IInjector';
 export { SimpleInjector } from './features/simple/SimpleInjector';
 export { IocInjector } from './features/ioc/IocInjector';
 export { IInjectMetadataCollector } from './features/ioc/IInjectMetadataCollector';
-export { IMethodsMetadataCollector } from './decorators/instanceHook/IMethodsMetadataCollector';
-export { MethodsMetadataCollector } from './decorators/instanceHook/MethodsMetadataCollector';
-export { createHookDecorator } from './decorators/instanceHook/decorators';
-export { createInjectDecorator, createInjectFnDecorator } from './decorators/ioc/decorators';
-export { InjectMetadataCollector } from './decorators/ioc/InjectMetadataCollector';
+export { IMethodsMetadataCollector } from './features/instanceHook/decorators/IMethodsMetadataCollector';
+export { MethodsMetadataCollector } from './features/instanceHook/decorators/MethodsMetadataCollector';
+export { createHookDecorator } from './features/instanceHook/decorators/decorators';
+export { createInjectDecorator, createInjectFnDecorator } from './features/ioc/decorators/decorators';
+export { InjectMetadataCollector } from './features/ioc/decorators/InjectMetadataCollector';
 export { InjectionToken } from './core/IServiceLocator';
 export { ProviderNotFoundError } from './errors/ProviderNotFoundError';
 export { MethodNotImplementedError } from './errors/MethodNotImplementedError';
